Fix some styling for togglebutton
